requirements.txt
requests==2.7.0
pandas==1.5.3
numpy==1.20.3
clickhouse_connect
psycopg2==2.9.5
apache-airflow==2.6.2